Move url handling code + fix litehtml url handling
- Move url handling code out of mainwindow
- Url handling for litehtml preview
- Clean up some textEdit related code and move it to textEdit
